Make g_thread_init dynamic.
This needs to define LIB_GTHREAD as the function is there.
Also g_thread_supported define is redefined for new glib version. There
is g_thread_get_initialized function for bindings but
as the whole g_thread_init is going away it makes no sense to add it in
the place of(conditionally) g_thread_supported.