untrusted setText and version checking
setText is for the most part more straightforward in CEF than in the other
browsers. Title events are fired correctly even in the absence of titles
and it appears to fire loading events correctly. However trusted
navigation is not currently working. The load_string method provided by
frame appears to refuse to run with the url set to a native url. If you
navigate manually to a native url first then attempt to set the text with
the url specified as the root it will not run. The code returns the error
-310, too many redirects.

As for the isInstalled, cef3 gives us the revision number. I just check if
it is leq the current one.