Adding JS dialog events
Implemented the JS Dialog handler. This means that there are now methods
for the 4 kinds of dialogs: confirm, alert, prompt and navigation. All of
which are styled after the windows Webkit dialogs of the same kind. The
methods take a JS Dialog Callback object that they execute after the
dialog is closed.

Caveats:
* The style of the before navigation dialog is different from the others.
  This is something I duplicated from the webkit implementation but it
  strikes me as odd.